+++ src/traffic_light/TrafficLight.h
@@ -3,12 +3,12 @@
 
 class TrafficLight {
   private:
-    int redPin;
-    int orangePin;
-    int greenPin;
+    uint8_t redPin;
+    uint8_t orangePin;
+    uint8_t greenPin;
 
   public:
-    TrafficLight(int r, int o, int g);
+    TrafficLight(uint8_t r, uint8_t o, uint8_t g);
     void begin();
     void apply(const Step& step);
 };
